Moving Job to be a top level class.
And making all derived Job classes hidden.
