+++ google/datalab/bigquery/__init__.py
@@ -17,10 +17,8 @@ from ._csv_options import CSVOptions
 from ._dataset import Dataset, Datasets
 from ._dialect import Dialect
 from ._federated_table import FederatedTable
-from ._job import Job
 from ._query import Query
 from ._query_output import QueryOutput
-from ._query_job import QueryJob
 from ._query_results_table import QueryResultsTable
 from ._query_stats import QueryStats
 from ._sampling import Sampling
