+++ tests/bigquery/query_tests.py
@@ -39,7 +39,7 @@ class TestCases(unittest.TestCase):
 
     sql = 'SELECT field1 FROM [table] LIMIT 1'
     q = TestCases._create_query(sql)
-    results = q.execute().results
+    results = q.execute().result()
 
     self.assertEqual(sql, results.sql)
     self.assertEqual('(%s)' % sql, q._repr_sql_())
@@ -60,7 +60,7 @@ class TestCases(unittest.TestCase):
     mock_api_insert_query.return_value = TestCases._create_insert_done_result()
 
     q = TestCases._create_query()
-    results = q.execute().results
+    results = q.execute().result()
 
     self.assertEqual(0, results.length)
 
@@ -79,7 +79,7 @@ class TestCases(unittest.TestCase):
     mock_api_insert_query.return_value = TestCases._create_incomplete_result()
 
     q = TestCases._create_query()
-    results = q.execute().results
+    results = q.execute().result()
 
     self.assertEqual(1, results.length)
     self.assertEqual('test_job', results.job_id)
@@ -91,7 +91,7 @@ class TestCases(unittest.TestCase):
     q = TestCases._create_query()
 
     with self.assertRaises(Exception) as error:
-      _ = q.execute().results
+      _ = q.execute().result()
     self.assertEqual('Unexpected response from server', str(error.exception))
 
   @staticmethod
